Text Animation Implementation
Added TextAnim class to create animated text
Added required functions and setup for it to work and not crash
Work in progress, the animation isn't as perfect as it should be
(spacing between the letter isn't perfectly handle at this point,
resulting in slight visual glitches when the text is move in the
completed string).
At this point only a letter by letter animation is supported.