Audio Hotfix 2
Readded the line that initialized the audio system in the engine, who
MUST be called before putting anything audio related in the
ressourceholder (otherwise the sound will come out like everything
except how it is suppose to, and also give warnings).
Placing it in the engine Init() is the best place to avoid any problem.