Audio system rework
->Cleaned up unnecessary functions from Audio.h/.cpp:
With our ResourceHolder, Audio shouldn't have to store any sound or
music within itself. It has all been cleaned up, resources are in the
ResourceHolder and we simply handle sound/music playback in Audio.
*Audio isn't it's own singleton anymore (You shouldn't have more than
one, ever), it's contained within engine, which is already a singleton
*Trashed a bunch of incomplete, redundant and now useless functions &
comments

->Adjusted volume level related functions:
Volume levels range from 0 to 128 but didn't handle case when volume was
between 100-128, so a rework of volume related functions was necessary.
Volume setting functions will still work using a range of 0-100 but that
range will represent a volume percentage, which is much easier to deal
with.

->Added a working example(SoundTest) to display the new audio changes
inside TestEnvironment. Instructions are in SoundTest.hpp

->Cleaned up remnants of sprite functions comments that were removed a
couple commits ago.